refactor(storage-ports): hoist 2 typed path helpers from root path_resolver (root-crate decomp)
Move `collection_data_path_typed` and `typed_identity_from_storage_assignment`
from `src/storage/trait_components/path_resolver.rs` into
`crates/storage/proximadb-storage-ports/src/path_resolver.rs` (where the
`CollectionPathResolver` trait + `StorageAssignment` already live from #902).
Re-export shims left at the old path so all callers
(`crate::storage::trait_components::path_resolver::*`) keep compiling — zero
diff to call sites.

Both functions' ONLY deps are foundation types, so the move is facade-free:
- `CollectionIdentity` (proximadb-kernel — dep added to storage-ports)
- `StoragePath` (proximadb-storage-common — already a dep)
- proto `StorageAssignment` (proximadb-proto — already a dep)

This is the LAST root-internal dep of `src/storage/traits` before it can move
to its own crate. The wal/index typed helpers stay root (no external caller
yet — no benefit to the move). Net +21 LOC (moved code + shim).
