+++ crates/storage/proximadb-storage-ports/Cargo.toml
@@ -19,6 +19,10 @@ async-trait.workspace = true
 serde.workspace = true
 proximadb-proto.workspace = true
 proximadb-storage-common.workspace = true
+# Foundation-tier: CollectionIdentity (typed path helpers) +
+# the StorageQuantizationEnginePort data types below, all foundation types
+# so the ports are facade-FREE (no DTO/adaptor layer).
+proximadb-kernel.workspace = true
 # Foundation-tier: the StorageQuantizationEnginePort returns these data types
 # (Vec<StorageQuantizedData>, &QuantizedVector, UnifiedQuantizationLevel) — moved to
 # foundation by #842, so the port is facade-FREE (no DTO/adaptor layer).
